Time: 791 ms (20.00%), Space: 90.9 MB (10.00%) - LeetHub

+++ 2233-maximum-product-after-k-increments/2233-maximum-product-after-k-increments.cpp
@@ -15,7 +15,7 @@ public:
             pq.push(a+1);
         }
         while(!pq.empty()){
-            mult = (long)(mult*pq.top())%mod;
+            mult = (long long)(mult*pq.top())%mod;
             pq.pop();
         }
         return mult;
